refactor(getfullgraph): head と base のフルグラフ生成処理を関数化
